requirements.txt
fastapi
uvicorn
mangum
pdfplumber
faiss-cpu
psycopg2-binary
transformers
openai
sentence-transformers
python-multipart
pydantic
numpy
pandas
pillow
requests
tqdm
flask
